+++ dominoes/assets/styles/calculator.js
@@ -1,0 +1,52 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#f5f5f5', 
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 20,
+  },
+  screen: {
+    width: '100%',
+    marginBottom: 20,
+    backgroundColor: 'white',
+    borderRadius: 10,
+    padding: 20,
+    elevation: 3,
+  },
+  input: {
+    fontSize: 36,
+    fontWeight: 'bold',
+    textAlign: 'right',
+    color: '#333',
+  },
+  buttonsContainer: {
+    width: '100%',
+  },
+  row: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginBottom: 10,
+  },
+  button: {
+    width: '22%',
+    height: 70,
+    backgroundColor: '#2196F3',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderRadius: 10,
+    elevation: 3,
+  },
+  buttonText: {
+    fontSize: 24,
+    color: 'white',
+    fontWeight: 'bold',
+  },
+  clearButton: {
+    backgroundColor: '#f44336',
+  },
+});
+
+export default styles;
