+++ dominoes/assets/styles/dashboard.js
@@ -1,0 +1,336 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#f5f5f5',
+  },
+  loadingContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#f5f5f5',
+  },
+  loadingText: {
+    marginTop: 10,
+    fontSize: 16,
+    color: '#333',
+  },
+  loginBox: {
+    width: '80%',
+    backgroundColor: 'white',
+    borderRadius: 10,
+    padding: 20,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 5,
+    alignSelf: 'center',
+    marginTop: '50%',
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginBottom: 20,
+    color: '#333',
+  },
+  inputContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderWidth: 1,
+    borderColor: '#ddd',
+    borderRadius: 8,
+    marginBottom: 15,
+    paddingHorizontal: 10,
+  },
+  icon: {
+    marginRight: 10,
+  },
+  input: {
+    flex: 1,
+    paddingVertical: 10,
+    fontSize: 16,
+    color: '#333',
+  },
+  eyeIcon: {
+    padding: 5,
+  },
+  loginButton: {
+    backgroundColor: '#3498db',
+    padding: 12,
+    borderRadius: 5,
+    alignItems: 'center',
+  },
+  loginButtonText: {
+    color: 'white',
+    fontWeight: 'bold',
+    fontSize: 16,
+  },
+  hamburgerButton: {
+    position: 'absolute',
+    top: 15,
+    left: 15,
+    zIndex: 50,
+  },
+  overlay: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: '#000',
+    zIndex: 90,
+  },
+  overlayTouchable: {
+    flex: 1,
+  },
+  sidebarContainer: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    bottom: 0,
+    zIndex: 100,
+  },
+  dashboardContainer: {
+    flexGrow: 1,
+    padding: 20,
+    paddingTop: 60,
+  },
+  dashboardTitle: {
+    fontSize: 28,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginBottom: 20,
+    color: '#333',
+  },
+  summaryCard: {
+    backgroundColor: 'white',
+    padding: 20,
+    borderRadius: 10,
+    marginBottom: 20,
+    borderWidth: 1,
+    borderColor: '#ddd',
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  summaryItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 15,
+  },
+  summaryText: {
+    fontSize: 16,
+    marginLeft: 10,
+    color: '#333',
+  },
+  section: {
+    marginVertical: 15,
+    width: '100%',
+  },
+  sectionTitle: {
+    fontSize: 20,
+    fontWeight: '600',
+    color: '#2196F3',
+    marginBottom: 15,
+    paddingLeft: 10,
+  },
+  playerCard: {
+    backgroundColor: '#FFFFFF',
+    borderRadius: 12,
+    padding: 15,
+    marginBottom: 12,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  playerHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 10,
+  },
+  playerRank: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    width: 30,
+    color: '#333',
+  },
+  playerInfo: {
+    flex: 1,
+    marginLeft: 10,
+  },
+  playerName: {
+    fontSize: 18,
+    fontWeight: '600',
+    color: '#333',
+  },
+  playerStats: {
+    fontSize: 14,
+    color: '#757575',
+  },
+  playerDetail: {
+    fontSize: 16,
+    color: '#757575',
+    marginBottom: 3,
+  },
+  comparisonContainer: {
+    backgroundColor: 'white',
+    padding: 15,
+    borderRadius: 10,
+    borderWidth: 1,
+    borderColor: '#ddd',
+  },
+  comparisonLegend: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 10,
+  },
+  legendItem: {
+    width: 20,
+    height: 10,
+    marginRight: 5,
+    marginLeft: 10,
+  },
+  winLegend: {
+    backgroundColor: '#4CAF50',
+  },
+  lossLegend: {
+    backgroundColor: '#F44336',
+  },
+  comparisonItem: {
+    marginBottom: 10,
+  },
+  comparisonName: {
+    fontSize: 16,
+    marginBottom: 5,
+    color: '#333',
+  },
+  barContainer: {
+    height: 20,
+    backgroundColor: '#f0f0f0',
+    borderRadius: 5,
+    overflow: 'hidden',
+  },
+  winBar: {
+    height: '50%',
+    backgroundColor: '#4CAF50',
+    position: 'absolute',
+    top: 0,
+  },
+  lossBar: {
+    height: '50%',
+    backgroundColor: '#F44336',
+    position: 'absolute',
+    bottom: 0,
+  },
+  actionButton: {
+    backgroundColor: '#2196F3',
+    padding: 15,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginBottom: 10,
+  },
+  actionButtonText: {
+    color: 'white',
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+  modalContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.5)',
+  },
+  modalContent: {
+    width: '80%',
+    backgroundColor: 'white',
+    borderRadius: 10,
+    padding: 20,
+    alignItems: 'center',
+  },
+  modalTitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 15,
+    color: '#333',
+  },
+  modalInput: {
+    width: '100%',
+    borderWidth: 1,
+    borderColor: '#ddd',
+    borderRadius: 8,
+    padding: 10,
+    marginBottom: 15,
+    fontSize: 16,
+  },
+  modalButtons: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: '100%',
+  },
+  modalButton: {
+    flex: 1,
+    padding: 10,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginHorizontal: 5,
+  },
+  cancelButton: {
+    backgroundColor: '#FF5252',
+  },
+  confirmButton: {
+    backgroundColor: '#4CAF50',
+  },
+  modalButtonText: {
+    color: 'white',
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+  netBadge: {
+    borderRadius: 12,
+    paddingVertical: 4,
+    paddingHorizontal: 10,
+  },
+  positiveNet: {
+    backgroundColor: '#E8F5E9',
+  },
+  negativeNet: {
+    backgroundColor: '#FFEBEE',
+  },
+  netText: {
+    fontWeight: 'bold',
+    fontSize: 14,
+  },
+  statsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginBottom: 8,
+  },
+  statItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  statValue: {
+    marginLeft: 5,
+    fontSize: 16,
+  },
+  gainText: {
+    color: '#4CAF50',
+  },
+  lossText: {
+    color: '#F44336',
+  },
+  progressBar: {
+    borderRadius: 2,
+    marginTop: 5,
+  },
+});
+
+export default styles;
