+++ dominoes/assets/styles/greating.js
@@ -1,0 +1,34 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  animation: {
+    width: 250,
+    height: 250,
+  },
+  title: {
+    fontSize: 32,
+    fontWeight: 'bold',
+    marginVertical: 20,
+    color: '#333',
+  },
+  button: {
+    backgroundColor: '#000',
+    paddingHorizontal: 30,
+    paddingVertical: 15,
+    borderRadius: 10,
+    marginBottom: 20,
+  },
+  buttonText: {
+    color: '#fff',
+    fontSize: 18,
+    fontWeight: '600',
+  },
+});
+
+export default styles;
