+++ dominoes/assets/styles/layout.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    position: 'relative'
+  },
+  content: {
+    flex: 1,
+    paddingBottom: 70 
+  }
+});
+
+export default styles;
