+++ dominoes/assets/styles/performanceChart.js
@@ -1,0 +1,131 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: '#f8f9fa',
+  },
+  header: {
+    marginBottom: 16,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: '#2c3e50',
+    marginBottom: 12,
+  },
+  dateFilterContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 8,
+  },
+  dateButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 4,
+    backgroundColor: '#fff',
+    paddingVertical: 6,
+    paddingHorizontal: 12,
+    borderRadius: 20,
+    borderWidth: 1,
+    borderColor: '#ddd',
+  },
+  dateText: {
+    fontSize: 14,
+    color: '#555',
+  },
+  dateSeparator: {
+    color: '#7f8c8d',
+    fontSize: 14,
+  },
+  playerCard: {
+    backgroundColor: '#fff',
+    borderRadius: 12,
+    padding: 16,
+    marginBottom: 12,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 2,
+  },
+  playerHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 12,
+  },
+  playerName: {
+    fontSize: 16,
+    fontWeight: '600',
+    color: '#34495e',
+    flex: 1,
+    marginRight: 8,
+  },
+  netBadge: {
+    paddingVertical: 4,
+    paddingHorizontal: 10,
+    borderRadius: 12,
+    minWidth: 80,
+    alignItems: 'center',
+  },
+  positiveNet: {
+    backgroundColor: 'rgba(76, 175, 80, 0.2)',
+  },
+  negativeNet: {
+    backgroundColor: 'rgba(244, 67, 54, 0.2)',
+  },
+  netText: {
+    fontWeight: 'bold',
+    fontSize: 14,
+  },
+  statsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginBottom: 12,
+  },
+  statItem: {
+    flex: 1,
+    alignItems: 'center',
+  },
+  statValue: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    marginTop: 4,
+  },
+  statSubtext: {
+    fontSize: 12,
+    color: '#7f8c8d',
+    marginTop: 2,
+  },
+  progressContainer: {
+    height: 6,
+    backgroundColor: 'rgba(244, 67, 54, 0.2)',
+    borderRadius: 3,
+    overflow: 'hidden',
+  },
+  progressBar: {
+    height: '100%',
+    backgroundColor: '#4CAF50',
+  },
+  transactionsContainer: {
+    marginTop: 12,
+    borderTopWidth: 1,
+    borderTopColor: '#eee',
+    paddingTop: 8,
+  },
+  transactionsTitle: {
+    fontWeight: '600',
+    fontSize: 12,
+    color: '#555',
+    marginBottom: 4,
+  },
+  transactionDetail: {
+    fontSize: 12,
+    color: '#666',
+    marginBottom: 2,
+  },
+});
+
+export default styles;
