+++ dominoes/assets/styles/sidebar.js
@@ -1,0 +1,40 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    width: 280,
+    backgroundColor: 'white',
+    height: '100%',
+    paddingTop: 50,
+    paddingHorizontal: 20,
+    shadowColor: '#000',
+    shadowOffset: { width: 2, height: 0 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 5,
+  },
+  closeButton: {
+    position: 'absolute',
+    top: 15,
+    right: 15,
+    padding: 5,
+  },
+  menuContainer: {
+    marginTop: 20,
+  },
+  menuItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 15,
+    borderBottomWidth: 1,
+    borderBottomColor: '#f0f0f0',
+  },
+  icon: {
+    marginRight: 15,
+  },
+  menuText: {
+    fontSize: 16,
+    color: '#333',
+  },
+});
+export default styles;
