+++ dominoes/src/components/Layout.js
@@ -1,6 +1,7 @@
 import { View, StyleSheet } from 'react-native';
 import Sidebar from './Sidebar';
 import Footer from './Footer';
+import styles from '../../assets/styles/layout';
 
 export default function Layout({ children, footerProps, sidebarProps }) {
   return (
@@ -16,13 +17,3 @@ export default function Layout({ children, footerProps, sidebarProps }) {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    position: 'relative'
-  },
-  content: {
-    flex: 1,
-    paddingBottom: 70 
-  }
-});
