+++ dominoes/src/screens/HomeScreen.js
@@ -17,6 +17,7 @@ import Sidebar from '../components/Sidebar';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import GameRepository from '../database/GameRepository';
 import PlayerRepository from '../database/PlayerRepository';
+import styles from '../../assets/styles/home';
 
 export default function HomeScreen({ route, navigation }) {
   const { players } = route.params;
@@ -418,215 +419,3 @@ export default function HomeScreen({ route, navigation }) {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flexGrow: 1,
-    padding: 20,
-    backgroundColor: '#f5f5f5',
-    marginTop: 50, 
-  },
-  hamburgerButton: {
-    position: 'absolute',
-    top: 15,
-    left: 15,
-    zIndex: 50,
-  },
-  overlay: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    backgroundColor: '#000',
-    zIndex: 90,
-  },
-  overlayTouchable: {
-    flex: 1,
-  },
-  sidebarContainer: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    bottom: 0,
-    zIndex: 100,
-  },
-  summaryContainer: {
-    flexGrow: 1,
-    padding: 20,
-    backgroundColor: '#f5f5f5',
-  },
-  summaryTitle: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: 30,
-    color: '#333',
-  },
-  summaryPlayerCard: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: 'white',
-    padding: 15,
-    borderRadius: 10,
-    marginBottom: 15,
-    borderWidth: 1,
-    borderColor: '#ddd',
-  },
-  summaryPosition: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#2196F3',
-    marginRight: 15,
-  },
-  summaryPlayerInfo: {
-    flex: 1,
-  },
-  summaryPlayerName: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#333',
-  },
-  summaryPlayerScore: {
-    fontSize: 16,
-    color: '#757575',
-  },
-  winnerBadge: {
-    color: '#4CAF50',
-    fontWeight: 'bold',
-    marginLeft: 10,
-  },
-  returnButton: {
-    backgroundColor: '#2196F3',
-    padding: 15,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginTop: 30,
-  },
-  returnButtonText: {
-    color: 'white',
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  turnIndicator: {
-    backgroundColor: '#2196F3',
-    padding: 15,
-    borderRadius: 10,
-    marginBottom: 20,
-    alignItems: 'center',
-  },
-  turnText: {
-    color: 'white',
-    fontSize: 16,
-  },
-  currentPlayer: {
-    color: 'white',
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginTop: 5,
-  },
-  playersContainer: {
-    marginBottom: 20,
-  },
-  playerCard: {
-    backgroundColor: 'white',
-    padding: 15,
-    borderRadius: 10,
-    marginBottom: 10,
-    borderWidth: 1,
-    borderColor: '#ddd',
-  },
-  activePlayerCard: {
-    borderColor: '#2196F3',
-    backgroundColor: '#e3f2fd',
-  },
-  winnerCard: {
-    borderColor: '#4CAF50',
-    backgroundColor: '#e8f5e9',
-  },
-  playerName: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#333',
-  },
-  playerScore: {
-    fontSize: 16,
-    color: '#2196F3',
-    marginTop: 5,
-  },
-  playerTurns: {
-    fontSize: 14,
-    color: '#757575',
-    marginTop: 3,
-  },
-  inputContainer: {
-    marginBottom: 15,
-  },
-  input: {
-    backgroundColor: 'white',
-    borderWidth: 1,
-    borderColor: '#ddd',
-    borderRadius: 8,
-    padding: 15,
-    fontSize: 16,
-  },
-  buttonsContainer: {
-    marginBottom: 15,
-  },
-  scoreButton: {
-    backgroundColor: '#BBDEFB',
-    padding: 15,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  currentPlayerButton: {
-    backgroundColor: '#2196F3',
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-  skipButton: {
-    backgroundColor: '#f44336',
-    padding: 15,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  skipButtonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: 'bold',
-    marginBottom: 30,
-  },
-  modalOverlay: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0,0,0,0.5)',
-  },
-  modalContent: {
-    width: '80%',
-    backgroundColor: 'white',
-    borderRadius: 10,
-    padding: 20,
-    alignItems: 'center',
-  },
-  modalText: {
-    fontSize: 18,
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  modalButton: {
-    backgroundColor: '#2196F3',
-    padding: 10,
-    borderRadius: 5,
-    width: '50%',
-    alignItems: 'center',
-  },
-  modalButtonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-});
