+++ dominoes/src/screens/PlayerSelectionScreen.js
@@ -14,6 +14,7 @@ import { fetchPlayersFromFirebase } from './../services/FirebaseService'
 import NetInfo from '@react-native-community/netinfo';
 import { database } from '../firebase/config';
 import { ref, get, push, remove } from 'firebase/database';
+import styles from '../../assets/styles/playerSelection';
 
 export default function PlayerSelectionScreen({ navigation }) {
   const [players, setPlayers] = useState([]);
@@ -392,139 +393,3 @@ useEffect(() => {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    padding: 20,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  playerRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginVertical: 5,
-  },
-  playerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 15,
-    flex: 1,
-    backgroundColor: '#f8f9fa',
-    borderRadius: 8,
-  },
-  checkbox: {
-    width: 24,
-    height: 24,
-    borderWidth: 2,
-    borderColor: '#2196F3',
-    borderRadius: 4,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  checked: {
-    width: 14,
-    height: 14,
-    backgroundColor: '#2196F3',
-    borderRadius: 2,
-  },
-  playerText: {
-    fontSize: 18,
-    marginLeft: 15,
-    color: '#333',
-  },
-  deleteButton: {
-    padding: 10,
-    marginLeft: 10,
-  },
-  deleteButtonText: {
-    fontSize: 24,
-    color: '#f44336',
-  },
-  addButton: {
-    backgroundColor: '#4CAF50',
-    padding: 15,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginVertical: 10,
-  },
-  startButton: {
-    backgroundColor: '#2196F3',
-    padding: 15,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginVertical: 10,
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0,0,0,0.5)',
-  },
-  modalContent: {
-    width: '80%',
-    backgroundColor: 'white',
-    borderRadius: 10,
-    padding: 20,
-  },
-  modalTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#ddd',
-    borderRadius: 6,
-    padding: 12,
-    marginBottom: 20,
-    fontSize: 16,
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  modalButton: {
-    flex: 1,
-    padding: 12,
-    borderRadius: 6,
-    alignItems: 'center',
-    marginHorizontal: 5,
-  },
-  cancelButton: {
-    backgroundColor: '#f44336',
-  },
-  confirmButton: {
-    backgroundColor: '#4CAF50',
-  },
-  modalButtonText: {
-    color: 'white',
-    fontWeight: 'bold',
-  },
-  refreshButton: {
-    backgroundColor: '#2196F3', 
-    padding: 10,
-    borderRadius: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginLeft: 10,
-  },
-  refreshText: {
-    color: '#FFFFFF',
-    fontWeight: 'bold',
-    fontSize: 14,
-  },
-
-});
